moved text color to separate file

+++ update.py
@@ -1,19 +1,33 @@
-from dav_utils_DAVIDE_PONZINI import commands, messages, user
+from dav_utils_DAVIDE_PONZINI import commands, messages, user, argument_parser, ArgumentAction
 
 user.require_root()
 
+argument_parser.add_argument('--update', help='check for updates', action=ArgumentAction.BOOLEAN_OPTIONAL, default=True)
+argument_parser.add_argument('--download-only', help='download updates but skip installation', action=ArgumentAction.STORE_TRUE)
+argument_parser.add_argument('--cleanup', help='removed unnecessary packages', action=ArgumentAction.BOOLEAN_OPTIONAL, default=True)
+
 try:
-    commands.execute('apt update')
-    messages.success('Downloaded package information')
+    if argument_parser.args.update:
+        commands.execute('apt update')
+        messages.success('Downloaded package information')
+    else:
+        messages.info('Skipping update')
 
-    commands.execute('apt dist-upgrade -y')
-    messages.success('Installed package upgrades')
+    if argument_parser.args.download_only:
+        commands.execute('apt dist-upgrade -y --download-only')
+        messages.success('Downloaded package upgrades')
+    else:
+        commands.execute('apt dist-upgrade -y')
+        messages.success('Installed package upgrades')
 
-    commands.execute('apt autoremove -y')
-    messages.success('Removed unused packages')
+    if argument_parser.args.cleanup:
+        commands.execute('apt autoremove -y')
+        messages.success('Removed unused packages')
 
-    commands.execute('apt autoclean')
-    messages.success('Cleaned up')
+        commands.execute('apt autoclean')
+        messages.success('Cleaned up')
+    else:
+        messages.info('Skipping cleanup')
 except commands.CalledProcessError as e:
     messages.error(e)
 except KeyboardInterrupt:
